Converted from old testing framework
