--- software/src/ASF/Makefile
@@ -1,9 +1,0 @@
-
-.PHONY: all
-all:
-
-.PHONY: clean
-clean:
-
-.PHONY: test
-test:
